requirements.txt
# This file was autogenerated via `uv export`.
-e .
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6 ; platform_system == 'Windows'
cryptography==45.0.5
deprecation==2.1.0
ecdsa==0.19.1
gitdb==4.0.12
gitpython==3.1.44
gotrue==2.12.3
greenlet==3.2.3 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markupsafe==3.0.2
narwhals==1.45.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.3.0
plotly==6.2.0
postgrest==1.1.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
realtime==2.5.3
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
storage3==0.12.0
streamlit==1.46.1
strenum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0 ; platform_system != 'Darwin'
websockets==15.0.1
